Files reorganized and getYouTubeIdFromInput function extracted.
