+++ src/YoutubeThumbnailEnhancer/noplay-hq.png
