+++ src/YoutubeThumbnailEnhancer/play-hq.png
