--- src/YoutubeThumbnailer.php
@@ -1,7 +1,0 @@
-<?php
-
-
-class YoutubeThumbnailer
-{
-
-}
